+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add multithreading, use `-c x`, where `x` is your desired bot amount
 - Move installation part of usage to own section in readme
 - Add on the run commands
+- Pause the current game
 
 ## [1.3.1] - 2017-08-23
 ### Changed